# Results from profiling slashing.BeginBlocker