--- a/slashing.1
+++ b/slashing.1
@@ -1,0 +1,6 @@
+goos: linux
+goarch: amd64
+pkg: github.com/cosmos/cosmos-sdk/x/slashing
+BenchmarkBeginBlocker100-8   	     100	  15535438 ns/op	 3915672 B/op	   49348 allocs/op
+PASS
+ok  	github.com/cosmos/cosmos-sdk/x/slashing	1.710s