--- a/slashing.2
+++ b/slashing.2
@@ -1,0 +1,6 @@
+goos: linux
+goarch: amd64
+pkg: github.com/cosmos/cosmos-sdk/x/slashing
+BenchmarkBeginBlocker100-8   	     100	  13009027 ns/op	 3654145 B/op	   43555 allocs/op
+PASS
+ok  	github.com/cosmos/cosmos-sdk/x/slashing	1.465s