--- a/x/slashing/abci.go
+++ b/x/slashing/abci.go
@@ -12,9 +12,9 @@ func BeginBlocker(ctx sdk.Context, req abci.RequestBeginBlock, k Keeper) {
 	// Iterate over all the validators which *should* have signed this block
 	// store whether or not they have actually signed it and slash/unbond any
 	// which have missed too many blocks in a row (downtime slashing)
-	swin := k.SignedBlocksWindow(ctx)
-	minswin := k.MinSignedPerWindow(ctx)
+	//swin := k.SignedBlocksWindow(ctx)
+	//minswin := k.MinSignedPerWindow(ctx)
 	for _, voteInfo := range req.LastCommitInfo.GetVotes() {
-		k.HandleValidatorSignatureReuse(ctx, voteInfo.Validator.Address, voteInfo.Validator.Power, voteInfo.SignedLastBlock, swin, minswin)
+		k.HandleValidatorSignatureOrig(ctx, voteInfo.Validator.Address, voteInfo.Validator.Power, voteInfo.SignedLastBlock)
 	}
 }