--- a/x/slashing/abci_test.go
+++ b/x/slashing/abci_test.go
@@ -1,15 +1,18 @@
 package slashing
 
 import (
+	"math/rand"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
 
 	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/crypto/ed25519"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	slashingkeeper "github.com/cosmos/cosmos-sdk/x/slashing/internal/keeper"
+	"github.com/cosmos/cosmos-sdk/x/slashing/internal/types"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 )
 
@@ -89,4 +92,69 @@ func TestBeginBlocker(t *testing.T) {
 	validator, found := sk.GetValidatorByConsAddr(ctx, sdk.GetConsAddress(pk))
 	require.True(t, found)
 	require.Equal(t, sdk.Unbonding, validator.GetStatus())
+}
+
+func benchmarkBeginBlocker(b *testing.B, num int) {
+	params := DefaultParams()
+	params.SignedBlocksWindow = 500
+	ctx, _, _, _, keeper, store, _ := slashingkeeper.CreateTestInputStore(b, params)
+
+	validators := []abci.Validator{}
+	signInfo := types.ValidatorSigningInfo{}
+	pubkey := [32]byte{}
+	for i := 0; i < num; i++ {
+		rand.Read(pubkey[:])
+		edkey := ed25519.PubKeyEd25519(pubkey)
+		addr := edkey.Address()
+		validators = append(validators, abci.Validator{Address: addr})
+		keeper.SetValidatorSigningInfo(ctx, sdk.ConsAddress(addr), signInfo)
+		keeper.AddPubkey(ctx, edkey)
+		// uncomment if MissedBlockBitArray is used
+		/*
+			for i := 0; i < 500; i++ {
+				keeper.SetValidatorMissedBlockBitArray(ctx, sdk.ConsAddress(addr), int64(i), true)
+			}
+		*/
+
+		keeper.StoreVoteArray(ctx, sdk.ConsAddress(addr), types.NewVoteArray(int(params.SignedBlocksWindow)))
+
+	}
+	store.Commit()
+	require.NoError(b, store.LoadVersion(0))
+	require.NoError(b, store.LoadVersion(1))
+	req := abci.RequestBeginBlock{
+		LastCommitInfo: abci.LastCommitInfo{
+			Votes: []abci.VoteInfo{},
+		},
+	}
+	op := abci.RequestBeginBlock{
+		LastCommitInfo: abci.LastCommitInfo{
+			Votes: []abci.VoteInfo{},
+		},
+	}
+
+	// two requests with opposite votes to trigger changes in missed block array
+	for i := range validators {
+		req.LastCommitInfo.Votes = append(req.LastCommitInfo.Votes, abci.VoteInfo{Validator: validators[i], SignedLastBlock: true})
+		op.LastCommitInfo.Votes = append(op.LastCommitInfo.Votes, abci.VoteInfo{Validator: validators[i]})
+	}
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ctx = ctx.WithBlockHeight(int64(i))
+		if i%2 == 0 {
+			BeginBlocker(ctx, req, keeper)
+		} else {
+			BeginBlocker(ctx, op, keeper)
+		}
+		commit := store.Commit()
+		// reset iavl tree with latest version
+		require.NoError(b, store.LoadVersion(0))
+		require.NoError(b, store.LoadVersion(commit.Version))
+	}
+
+}
+
+func BenchmarkBeginBlocker100(b *testing.B) {
+	benchmarkBeginBlocker(b, 100)
 }