results related to https://github.com/cosmos/cosmos-sdk/issues/4977

I found right knobs and made slashing.BeginBlocker work in a similar way how it works in real scenarios, it is somewhat hacky, but profile picture resembles one reported by @alexanderbez 

As it was noted in one of the issues, majority of time is spent due to read amplifications in iavl store. I tried approach similar  to one proposed by @ValarDragon , but i kept it simple and instead of partitioning bit array i basically created one per validator. Considered that array is limited by the size of sliding window (10000 bits or even bigger?) and we need only one per validator - this is >~1 mb per 1000 validators. I also collected db sizes after benchmark, with constant number of iterations. and storing single bitarray is clearly beneficial, unless i missed something :)

Before proceeding with this pr i will try to sync mainnet with this change (tests are passing so i assume nothing is broken) and see if this change has an expected effect.
Until then this is mostly for any early feedback, as it is highly appreciated.

Pasting results from profile.md, and there are two weblists for handleValidatorSignature attached to this PR, one before changes and one after.

slashing profiling results
==========================

original bench
===

benchmark
---

```
goos: linux
goarch: amd64
pkg: github.com/cosmos/cosmos-sdk/x/slashing
BenchmarkBeginBlocker100-8           100          30795066 ns/op         7486165 B/op      90541 allocs/op
PASS
ok      github.com/cosmos/cosmos-sdk/x/slashing	5.915s
```

db size
---

```
14M	/tmp/bench-slashing111093085/test.db
14M	/tmp/bench-slashing111093085
```

missed array stored as a single batch
===

benchcmp to original
---
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkBeginBlocker100-8     30795066      15535438      -49.55%

benchmark                      old allocs     new allocs     delta
BenchmarkBeginBlocker100-8     90541          49348          -45.50%

benchmark                      old bytes     new bytes     delta
BenchmarkBeginBlocker100-8     7486165       3915672       -47.69%
```

db size
---

```
2,3M	/tmp/bench-slashing347257969/test.db
2,3M	/tmp/bench-slashing347257969
```


reuse SignedBlocksWindow and MinSignedBlocksWindow
===

benchcmp to original
---

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkBeginBlocker100-8     30795066      13009027      -57.76%

benchmark                      old allocs     new allocs     delta
BenchmarkBeginBlocker100-8     90541          43555          -51.89%

benchmark                      old bytes     new bytes     delta
BenchmarkBeginBlocker100-8     7486165       3654145       -51.19%
```
______

For contributor use:

- [ ] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [ ] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added  relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the Github PR explorer

______

For admin use:

- [ ] Added appropriate labels to PR (ex. `WIP`, `R4R`, `docs`, etc)
- [ ] Reviewers assigned
- [ ] Squashed all commits, uses message "Merge pull request #XYZ: [title]" ([coding standards](https://github.com/tendermint/coding/blob/master/README.md#merging-a-pr))